[heap][update] Skip KGs not belonging to this heap state backend on restore
Motivation:
in new incremental mode, heap backend wraps KeyGroupsStateHandle
into IncrementalRemoteKeyedStateHandle. The latter don't compute
intersection because it is not directly aware of the offsets. So it just
returns a full keyrange if there is SOME intersection.
On recovery, unused keyGroups are filtered out by RocksDB state backend.
With this change, Heap state backend does the same.