[FLINK-24611] Adjust shared state entry de-duplication
Because shared state is now registered before checkpoint completion
it is possible that a new state should be kept in case of collision.

For example, if a task sent state for the first time and then failed
without completing checkpoint and restarting JM; then that old state
should be discarded in case of collision by key.