[FLINK-25987][state/changelog] Replace lastSqn with nextSqn()
The current semantics of ChangelogWriter.lastAppendedSqn doesn't allow
to distinguish between initial empty and non-empty states.
This results in wrong SQN passed to truncate and persist methods.

This change solves this by replacing lastAppendedSqn with nextSqn.
It essentially moves the responsibility to call sqn.next() on materialization
from backend to writer which has the knowledge of whether there were any changes or not.