[hotfix][state/heap] Fix race condition in CopyOnWriteStateMap
Currently, handling chained entry on copy-on-write uses object
identity to find the wanted entry in the chain. However, if
the same method is running concurrently, the object in the chain
can be replaced by its copy; the condition will never be met and
the chain end will be reached, causing an NPE.

With some tiny changes in timings (i.e. overriding methods of
CopyOnWriteStateMap), StateBackendTestBase.testValueStateRace
fails when running repeatedly (~4 out of 100 runs).

This change replaces object identity with key+namespace equality
in the condition.

The overhead should not be significant because the same check is
already performed to find the element before copying.