[hotfix][state/heap] Bump CopyOnWriteStateMap entry version before write
CopyOnWriteStateMap copies the entry before returning it to the client
for update. This also updates its state and entry versions.

However, if the entry is NOT used by any snapshots, the versions
will stay the same despite that state is going to be updated.

With incremental checkpoints, this causes such updated version to
be ignored in the next snapshot.

This change bumps the state version in this case (entry version
stays the same).